+++ Cargo.toml
@@ -9,9 +9,6 @@ members = [
     "crates/vrf",
     "tee_vlc",
     "tee_llm", 
-    "operator/node_api",
-    "operator/db_sql",
-    "operator/runer",
     "worker",
 ]
 
@@ -48,7 +45,7 @@ inherits = "release"
 [workspace.dependencies]
 futures = "0.3.30"
 tokio = "1.36.0"
-tracing-subscriber = "0.3.18"
+tracing-subscriber = {version = "0.3.18", features = ["env-filter"]}
 anyhow = { version = "1.0.79", features = ["backtrace"] }
 actix-web = "4.8.0"
 serde = { version = "1.0.195", features = ["derive"] }
@@ -61,4 +58,3 @@ bincode = "1.3.3"
 nix = { version = "0.28.0", features = ["socket", "sched", "resource"] }
 
 tee_llm = { path = "./tee_llm", features = ["nitro-enclaves"]}
-operator-runer = { path = "./operator/runer"}
