--- operator/README.md
@@ -1,11 +1,0 @@
-# Aos Operator
-
-### Run operator
-
-```shell
-cargo build --features nitro-enclaves --release
-
-./target/release/operator-runer -i postgres://postgres:hetu@0.0.0.0:5432/operator_db
-
-./target/release/operator-runer  -c ./docs/template/config-operator.yaml
-```
