--- operator/db_sql/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "db_sql"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-parking_lot = "0.12.3"
-once_cell = "1.4.1"
-thiserror = "1.0.58"
-derive_more = "0.99"
-serde_json = { version = "1" }
-async-trait = "0.1.79"
-tracing = "0.1.40"
-shrinkwraprs = "0.3.0"
-futures = "0.3.28"
-tokio = { version = "1", features = [
-    "net",
-    "rt-multi-thread",
-    "macros",
-    "time",
-] }
-sea-orm = { version = "0.12.15", features = [ 
-    "sqlx-postgres", 
-    "sqlx-sqlite", 
-    # "runtime-tokio-native-tls",
-    "runtime-async-std-native-tls",
-    "macros" 
-] }
-sea-orm-migration = "0.12.15"
-async-std="1.12.0"
-chrono = "0.4.38"
-common ={ path = "../../crates/common", version = "0.1.0"}
