--- operator/db_sql/src/error.rs
@@ -1,8 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum DatabaseError {
-    //todo add error type
-}
-
-pub type DatabaseResult<T> = Result<T, DatabaseError>;
