--- operator/db_sql/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod error;
-pub mod pg;
