--- operator/db_sql/src/pg/entities/clock_infos.rs
@@ -1,24 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "clock_infos")]
-pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: i64,
-    pub clock: String,
-    #[sea_orm(unique)]
-    pub clock_hash: String,
-    pub node_id: String,
-    pub message_id: String,
-    #[sea_orm(column_type = "Binary(BlobSize::Blob(None))")]
-    pub raw_message: Vec<u8>,
-    pub event_count: i64,
-    pub create_at: Option<DateTime>,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {}
-
-impl ActiveModelBehavior for ActiveModel {}
