--- operator/db_sql/src/pg/entities/mod.rs
@@ -1,5 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
-
-pub mod prelude;
-
-pub mod clock_infos;
