--- operator/db_sql/src/pg/entities/prelude.rs
@@ -1,3 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
-
-pub use super::clock_infos::Entity as ClockInfos;
