--- operator/db_sql/src/pg/mod.rs
@@ -1,3 +1,0 @@
-pub mod entities;
-pub mod pg_client;
-pub mod migrator;
