--- operator/node_api/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "node_api"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0.22"
-serde_yaml = "0.9"
-actix-web = "4.8.0"
-eyre = "0.6.12"
-alloy-primitives = "0.7.2"
-serde = { version = "1.0", features = [ "derive" ] }
-reqwest = { version = "0.12.4", features = ["json", "multipart"] }
-tools = { version = "0.1.0", path = "../../crates/tools" }
-tee_llm ={version = "0.1.0", path = "../../tee_llm" }
-tokio = { version = "1.35.1", features = ["net", "time", "sync", "rt", "signal", "macros", "rt-multi-thread", "fs", "process", "io-util"] }
