--- operator/node_api/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod config;
-pub mod error;
-
