--- operator/runer/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "operator-runer"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-common = { path = "../../crates/common"}
-tools = { version = "0.1.0", path = "../../crates/tools" }
-node_api = {version ="0.1.0", path = "../node_api" }
-db_sql ={version = "0.1.0", path = "../db_sql" }
-tee_llm ={version = "0.1.0", path = "../../tee_llm" }
-alloy-wrapper = { path = "../../crates/alloy-wrapper"}
-structopt = "0.3.11"
-tracing = "0.1.40"
-base64 = "0.13"
-tracing-subscriber = "0.3.18"
-tokio = { version = "1.36.0", features = ["full"] }
-sea-orm = { version = "0.12.15", features = [ 
-    "sqlx-postgres", 
-    "sqlx-sqlite", 
-    "runtime-async-std-native-tls",
-    "macros" 
-] }
-alloy-primitives = "0.7.7"
-alloy = { version = "0.2.0"}
-serde = { version = "1", features = ["derive"] }
-serde_json = { version = "1" }
-url = "2.5.0"
-sha2 = "0.10.8"
-chrono = "0.4"
-hex = "0.4.3"
-reqwest = { version = "0.12.4", features = ["json", "multipart"] }
-actix-web = "4.8.0"
-
-[lints]
-workspace = true
