--- operator/runer/src/cli/mod.rs
@@ -1,2 +1,0 @@
-pub mod operator;
-pub mod command;
