--- operator/runer/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod operator;
-pub mod storage;
-pub mod node_factory;
-pub mod handler;
-pub mod api;
-pub mod cli;
